add tests, rollup, fix problems

+++ src/main.rs
@@ -1,20 +1,24 @@
 mod templates;
 mod wasm_opt;
-
 use cargo_metadata::{self, Error, Metadata, Package};
-use std::{collections::BTreeSet, path::PathBuf, process::Command, str::FromStr, sync::Arc};
+use std::{
+    collections::BTreeSet, io::Write, path::PathBuf, process::Command, str::FromStr, sync::Arc,
+};
 use structopt::StructOpt;
+use wasm_opt::WasmOpt;
 
 const WASM32_UNKNOWN_UNKNOWN: &str = "wasm32-unknown-unknown";
 const WASM_BINDGEN: &str = "wasm-bindgen";
 const WASM_BINDGEN_CLI: &str = "wasm-bindgen-cli";
 const OUT_DIR: &str = "dist/js";
 
-fn path_to_cli(wasm_bindgen_version: &str) -> String {
-    format!(
-        "./target/{}/{}/bin/{}",
-        WASM_BINDGEN_CLI, wasm_bindgen_version, WASM_BINDGEN
-    )
+fn path_to_cli(wasm_bindgen_version: &str) -> PathBuf {
+    let mut path = PathBuf::from("target");
+    path.push(WASM_BINDGEN_CLI);
+    path.push(wasm_bindgen_version);
+    path.push("bin");
+    path.push(WASM_BINDGEN);
+    path
 }
 
 #[derive(Clone)]
@@ -24,19 +28,43 @@ impl Cargo {
     // TODO: Only run if not there... `exists()` still works, if just the file is deleted?
     // if !Path::new(&path_to_cli(wasm_bindgen_version)).exists() {
     // }
+    // TODO: Write test to make sure that correct version is installed: check .crates.toml file
     fn install_wasm_bindgen_cli(&self, wasm_bindgen_version: &str) {
-        let path = format!("./target/{}/{}", WASM_BINDGEN_CLI, wasm_bindgen_version);
-        println!("Installing {}: {}", WASM_BINDGEN_CLI, wasm_bindgen_version);
-        let mut cmd = Command::new(&self.0);
-        cmd.args(&["install", "--root", &path, "--", WASM_BINDGEN_CLI]);
-        // TODO: Show this so can see something is happening?
-        cmd.output().expect("Unable to install wasm_bindgen_cli");
+        let mut path = path_to_cli(wasm_bindgen_version);
+        if path.exists() {
+            println!(
+                "{} {} already installed",
+                WASM_BINDGEN_CLI, wasm_bindgen_version
+            );
+        } else {
+            path.pop();
+            path.pop();
+            println!("Installing {}: {}", WASM_BINDGEN_CLI, wasm_bindgen_version);
+            let mut cmd = Command::new(&self.0);
+            cmd.args(&[
+                "install",
+                "--root",
+                &path.display().to_string(),
+                "--version",
+                wasm_bindgen_version,
+                "--",
+                WASM_BINDGEN_CLI,
+            ]);
+            // TODO: Show this so can see something is happening?
+            cmd.output().expect("Unable to install wasm_bindgen_cli");
+        }
     }
 
-    fn build_wasm32_unkwown_unknown(&self, package_name: &str, opt: &Opt) {
+    fn build_wasm32_unknown_unknown(&self, package_name: &str, opt: &Opt) {
         println!("Building {} for {}", WASM32_UNKNOWN_UNKNOWN, package_name);
         let mut cmd = Command::new(&self.0);
-        cmd.args(&["build", "--target", WASM32_UNKNOWN_UNKNOWN]);
+        cmd.args(&[
+            "build",
+            "--package",
+            package_name,
+            "--target",
+            WASM32_UNKNOWN_UNKNOWN,
+        ]);
         if opt.release {
             cmd.arg("--release");
         }
@@ -141,31 +169,25 @@ impl PackageInfo {
             cmd.arg("--no-demangle");
         }
 
-        let mut out_dir = PathBuf::new();
-        out_dir.push(opt.out_dir.as_deref().unwrap_or_else(|| OUT_DIR));
-        cmd.args(&["--out-dir", out_dir.to_str().unwrap()]);
+        let mut out_wasm = PathBuf::new();
+        out_wasm.push(opt.out_dir.as_deref().unwrap_or_else(|| OUT_DIR));
+        cmd.args(&["--out-dir", &out_wasm.display().to_string()]);
 
-        println!(
-            "Building js glue code for {} with wasm-bindgen = {}. Output at: {}",
-            self.get_package_name(),
-            self.wasm_bindgen_version,
-            out_dir.display()
-        );
-        let _ = cmd.output().expect("unable to build js glue code");
-        // std::io::stdout().write_all(&output.stdout).unwrap();
-        // std::io::stderr().write_all(&output.stderr).unwrap();
+        println!("Building js glue code for {}", self.get_package_name(),);
 
-        let output_wasm = format!("{}/{}_bg.wasm", out_dir.display(), self.get_package_name());
-        if opt.wasm_opt {
-            match wasm_opt::WasmOpt::new() {
-                Some(wasm_opt) => wasm_opt.run(&output_wasm, opt),
-                None => eprint!("Could not install wasm-opt"),
+        let output = cmd.output().expect("unable to build js glue code");
+        std::io::stdout().write_all(&output.stdout).unwrap();
+        std::io::stderr().write_all(&output.stderr).unwrap();
+
+        out_wasm.push(format!("{}_bg.wasm", self.get_package_name()));
+        if let Some(SubCmds::WASM_OPT(wasm_opt)) = &opt.subs {
+            if let Err(e) = wasm_opt.try_install_and_run(&out_wasm, opt) {
+                eprint!("Error running wasm-opt: {}", e)
             }
         }
     }
 }
 
-//TODO: What if workspace is a crate in of itself?
 #[derive(Clone)]
 struct BindgenPackages {
     packages: Vec<PackageInfo>,
@@ -187,10 +209,10 @@ impl BindgenPackages {
         Ok(BindgenPackages { packages, cargo })
     }
 
-    fn build_wasm32_unkwown_unknown(&self, opt: &Opt) {
+    fn build_wasm32_unknown_unknown(&self, opt: &Opt) {
         for p in &self.packages {
             self.cargo
-                .build_wasm32_unkwown_unknown(&p.get_package_name(), opt)
+                .build_wasm32_unknown_unknown(&p.get_package_name(), opt)
         }
     }
 
@@ -216,9 +238,14 @@ impl BindgenPackages {
             println!("Cleaning out-dir: {}", &out_dir.display());
             std::fs::remove_dir_all(&out_dir).unwrap_or(());
         }
-
         for p in &self.packages {
             p.build_wasm_js(opt)
+        }
+        // TODO: Here should look for a rollup.config.js file & run it?
+        if let Some(WasmTarget::Rollup) = opt.target {
+            let bootstrap = templates::rollup_bootstrap_js(self.packages.as_slice(), &out_dir);
+            out_dir.push("bootstrap.js");
+            std::fs::write(out_dir, bootstrap).unwrap()
         }
     }
 }
@@ -227,7 +254,10 @@ impl BindgenPackages {
 #[derive(StructOpt)]
 enum WasmTarget {
     Web,
-    Bundler,
+    Rollup,
+    // NoModules,
+    /// For use with
+    Webpack,
     // Nodejs,
     // Deno,
 }
@@ -237,11 +267,13 @@ impl FromStr for WasmTarget {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s.to_ascii_lowercase().as_str() {
             "web" => Ok(WasmTarget::Web),
-            "bundler" => Ok(WasmTarget::Bundler),
+            "webpack" => Ok(WasmTarget::Webpack),
+            "rollup" => Ok(WasmTarget::Rollup),
+            // "no-modules" => Ok(WasmTarget::NoModules),
             // "nodejs" => Ok(WasmTarget::Nodejs),
             // "deno" => Ok(WasmTarget::Deno),
             _ => Err(format!(
-                "'{}' is not an allowed target. Supported options are: web (default), bundler", //, nodejs, deno",
+                "'{}' is not an allowed target. Supported options are: web (default), webpack", //, nodejs, deno",
                 s
             )),
         }
@@ -252,7 +284,9 @@ impl AsRef<str> for WasmTarget {
     fn as_ref(&self) -> &str {
         match self {
             WasmTarget::Web => "web",
-            WasmTarget::Bundler => "bundler",
+            WasmTarget::Rollup => "web",
+            // WasmTarget::NoModules => "no-modules",
+            WasmTarget::Webpack => "bundler",
             // WasmTarget::Nodejs => "nodejs",
             // WasmTarget::Deno => "deno",
         }
@@ -272,7 +306,7 @@ struct Opt {
     #[structopt(long, short)]
     typescript: bool,
 
-    /// Target to compile the js glue code to: web (default), bundler
+    /// Target to compile the js glue code to: web (default), webpack
     #[structopt(long)]
     target: Option<WasmTarget>,
 
@@ -303,9 +337,25 @@ struct Opt {
     #[structopt(long)]
     no_demangle: bool,
 
-    /// Runs wasm-opt https://github.com/WebAssembly/binaryen#wasm-opt
-    #[structopt(long)]
-    wasm_opt: bool,
+    #[structopt(subcommand)]
+    subs: Option<SubCmds>,
+}
+
+// TODO: Is giant struct a good idea... or should just use a vec? Use a macro to parse the wasm-opt.txt file...
+
+#[derive(StructOpt, Debug)]
+#[allow(non_camel_case_types)]
+enum SubCmds {
+    /// Runs wasm-opt https://github.com/WebAssembly/binaryen#wasm-opt.
+    /// Its runs as `wasm-opt path_to_wasm --o path_to_wasm [FLAGS]`
+    /// and uses a default optimization of '--O'.
+    /// It will pass through any other relevant flags that are parsed to wasm build: --enable-reference-types
+    WASM_OPT(WasmOpt),
+}
+impl Default for SubCmds {
+    fn default() -> Self {
+        Self::WASM_OPT(WasmOpt::default())
+    }
 }
 
 // TODO: Static version of rollup.js for packaging?
@@ -346,7 +396,7 @@ impl CargoWasm {
         let handler = std::thread::spawn(move || {
             bp.install_wasm_bindgen_cli();
         });
-        bindgen_packages.build_wasm32_unkwown_unknown(opt);
+        bindgen_packages.build_wasm32_unknown_unknown(opt);
         handler.join().unwrap();
         bindgen_packages.build_wasm_js(opt);
     }
@@ -360,5 +410,37 @@ fn main() {
     match std::env::var("CARGO") {
         Ok(cargo) => cargo_wasm.run(Cargo(cargo)),
         Err(e) => eprintln!("{}", e),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn installs_correct_bindgen_version() {
+        let crates = ".crates.toml";
+
+        let v_1 = "0.2.68";
+        let cargo_1 = Cargo(std::env::var("CARGO").unwrap());
+        let mut path_1 = PathBuf::from("target/wasm-bindgen-cli");
+        path_1.push(v_1);
+        path_1.push(crates);
+
+        let v_2 = "0.2.67";
+        let cargo_2 = cargo_1.clone();
+        let mut path_2 = PathBuf::from("target/wasm-bindgen-cli");
+        path_2.push(v_2);
+        path_2.push(crates);
+
+        let handler = std::thread::spawn(move || {
+            cargo_1.install_wasm_bindgen_cli("0.2.68");
+            let file_1 = std::fs::read_to_string(path_1).unwrap();
+            assert!(file_1.contains(v_1))
+        });
+
+        cargo_2.install_wasm_bindgen_cli("0.2.67");
+        let file_2 = std::fs::read_to_string(path_2).unwrap();
+        assert!(file_2.contains(v_2));
+        handler.join().unwrap();
     }
 }
